requirements.txt
transformers>=4.30
torch>=2.0
python-dotenv>=1.0
matplotlib>=3.5
gradio>=4.0
groq>=0.4.0
